refactor: extract sessionruntime execution-target resolution
Move SyncExecutionTarget, worktree reminder persistence/normalization, and the
store/claim/execution-target resolvers out of service.go into execution_target.go.
